pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py==2.5.2",
    "httpx==0.28.1",
    "pymongo==4.13.2",
    "python-dotenv==1.1.0",
    "typing-extensions==4.13.2",
    "ua-generator==2.0.7",
]
